requirements.txt
opencv-python==4.7.0.72
numpy==1.23.5
ultralytics==8.0.114
pyserial